+++ web/src/state_data.ts
@@ -21,20 +21,6 @@ export const narrow_term_schema = z.object({
 export type NarrowTerm = z.output<typeof narrow_term_schema>;
 // Sync this with zerver.lib.events.do_events_register.
 
-const one_time_notice_schema = z.object({
-    name: z.string(),
-    type: z.literal("one_time_notice"),
-});
-
-/* We may introduce onboarding step of types other than 'one time notice'
-in future. Earlier, we had 'hotspot' and 'one time notice' as the two
-types. We can simply do:
-const onboarding_step_schema = z.union([one_time_notice_schema, other_type_schema]);
-to avoid major refactoring when new type is introduced in the future. */
-const onboarding_step_schema = one_time_notice_schema;
-
-export type OnboardingStep = z.output<typeof onboarding_step_schema>;
-
 export const custom_profile_field_schema = z.object({
     display_in_profile_summary: z.optional(z.boolean()),
     field_data: z.string(),
@@ -57,7 +43,6 @@ export const current_user_schema = z.object({
     is_guest: z.boolean(),
     is_moderator: z.boolean(),
     is_owner: z.boolean(),
-    onboarding_steps: z.array(onboarding_step_schema),
     user_id: z.number(),
 });
 // Sync this with zerver.lib.events.do_events_register.
